Fix when native .wait is called but not assigned, parent function that awaits is won't be resolved. And add --keep-temp command line options

+++ bin/compile.js
@@ -100,6 +100,7 @@ if (process.argv[2] && process.argv[2].indexOf('.py') !== -1) {
       });
       fs.writeFileSync(path.join(process.cwd(), file + '.wasm.map'), JSON.stringify(source_map));
     }
+    if (process.argv.indexOf('--keep-temp') !== -1) process.exit();
     try {
       fs.rmdirSync(tempdir, {recursive: true});
     }
